Bring other options onboard ..
.. the new CodeStyle options UI page.

Notes:
1. UseVarForDeclarations - the option used by some features isn't
onboard, it is to be replaced by the new use var options.

2. Since the existing options are part of public api, we keep the
options' data structure as such and map the type to a corresponding view
which is used to update the UI. Since they were all underlying boolean
options, we introduce a BooleanCodeStyleOptionViewModel that maps to a
CodeStyleOption type and back.

3. update the StyleViewModel to create options using these new view
models.