Show goto tools options for codestyle features
ToDo: still haven't done it for use var options since they're defined in
CSharp Features and it is not yet referenced by ServicesVisualStudio.
The other options are in Workspaces layer. I have to think about moving
the options to workspaces so that they work with the simplifier, instead
of simply adding a reference here ? Also, I have to make them
PerLanguage options.?