support Brush b = SystemBrushes.Background scenario
This is based on a heuristic. I'm afraid this check is very generic and
may result in false positive matches. So, run an experiment to verify
correctness: Test this heuristic on some framework assembly, say the wpf
one that has colors and brushes and see how many false positives we get.